requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.2
django-cors-headers==2.1.0
djangorestframework==3.6.3
dlib==19.8.1
face-recognition==1.0.0
face-recognition-models==0.3.0
gunicorn==19.6.0
mysqlclient==1.3.12
numpy==1.13.3
Pillow==5.0.0
psycopg2==2.7.3.2
python-decouple==3.1
scipy==1.0.0
static3==0.7.0
Unipath==1.0
whitenoise==3.3.1
